update db definition for deal proposals

+++ include/sql/ps_filecoin_schema.sql
@@ -88,7 +88,7 @@ create table if not exists deal_proposals
     "endEpoch"             integer,
     "clientCollateral"     numeric,
     "providerCollateral"   numeric,
-    "storagePricePerEpoch" integer,
+    "storagePricePerEpoch" numeric,
     label                  varchar,
     verified               boolean
 );
